Factor out frontend workflow functions into util
