+++ frontend/src/components/TopPanel.tsx
@@ -11,8 +11,8 @@ import { Send } from "lucide-react";
 import { Tooltip, TooltipContent, TooltipTrigger } from "./ui/tooltip";
 import GeneratedWorkflowDialog from "./GeneratedWorkflowDialog";
 import { ActionNode } from "./nodes/ActionNode";
-import { generateId } from "@/lib/utils";
-import { Edge, Step, Workflow } from "@/types/workflow";
+import { Workflow } from "@/types/workflow";
+import { makeWorkflow } from "@/lib/workflowUtils";
 
 export default function TopPanel({
   workflowName,
@@ -64,7 +64,6 @@ export default function TopPanel({
       return;
     }
 
-    const steps: Step[] = [];
     let currNodeId = nodeConnectionGraph.get("trigger");
     while (currNodeId !== undefined) {
       const node = getNode(currNodeId);
@@ -92,45 +91,11 @@ export default function TopPanel({
         setSubmitting(false);
         return;
       }
-
-      const newStep: Step = {
-        id: node.id,
-        name: data.definition.name,
-        position: node.position,
-        step_id: data.definition.id,
-        inputs: data.inputs,
-      };
-      steps.push(newStep);
       currNodeId = nodeConnectionGraph.get(currNodeId);
     }
 
     const triggerNode = getNode("trigger") as TriggerNode;
-    const wfRequest: Workflow = {
-      name: workflowName,
-      run_name: workflowName,
-      trigger: {
-        id: triggerNode.id,
-        position: triggerNode.position,
-        conditions: [
-          {
-            event: triggerNode.data.trigger,
-            config: {},
-          },
-        ],
-      },
-      jobs: [
-        {
-          id: generateId(),
-          name: "Main Job",
-          steps: steps,
-          step_edges: edges.map(
-            (e) =>
-              ({ id: e.id, source: e.source, target: e.target } satisfies Edge)
-          ),
-        },
-      ],
-      job_edges: [],
-    };
+    const wfRequest = makeWorkflow(workflowName, triggerNode, nodes, edges);
 
     const newWorkflow = await fetcher<Workflow>("/workflows", {
       method: "POST",
